requirements.txt
torch>=2.0.0
torchvision
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.6.0
Pillow>=9.0.0
openai>=1.0.0
python-dotenv>=1.0.0
huggingface-hub>=0.16.0
tqdm>=4.64.0
